refactor : import 경로 수정 , 텍스트 필드 글자 색 수정
